requirements.txt
playwright
python-dotenv
requests
google-generativeai
